requirements.txt
flask
openai
python-dotenv
gunicorn
flask-cors
supabase